+++ internal/repository/document_processing_job_interface.go
@@ -8,7 +8,7 @@ import (
 
 // DocumentProcessingJobRepository 定义文档处理任务数据访问能力
 type DocumentProcessingJobRepository interface {
-	CreateProcessJob(ctx context.Context, job *entity.DocumentProcessingJob, allowedDocumentStatuses []int16, processingDocumentStatus int16) (bool, error)
+	CreateProcessJob(ctx context.Context, job *entity.DocumentProcessingJob, allowedDocumentStatuses []int, processingDocumentStatus int) (bool, error)
 	ListByDocument(ctx context.Context, userID, documentID string) ([]entity.DocumentProcessingJob, error)
 	FindByID(ctx context.Context, userID, jobID string) (entity.DocumentProcessingJob, bool, error)
 }
